- moved Generation of AnalyticsSourceDate data out of the block and into the service layer so that it can be run without the block (for example, in a migration)
